requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
shap>=0.41.0
imbalanced-learn>=0.8.0
jupyter>=1.0.0
notebook>=6.4.0
plotly>=5.0.0
xgboost>=1.5.0
lightgbm>=3.3.0
catboost>=1.0.0
optuna>=2.10.0
mlflow>=1.20.0
pytest>=6.2.0
black>=21.0.0
flake8>=3.9.0
PyYAML>=6.0
